fix: Multiple Conflicting App Entry Points
