+++ pages/__init__.py
@@ -1,1 +1,9 @@
-# Pages module for LegalEase AI Streamlit multi-page app
+"""
+Pages module for LegalEase AI Streamlit multi-page app
+
+This directory contains the individual pages:
+- 0_Home.py: Judgment analysis (main feature)
+- 1_Deadlines.py: Case deadline management
+- 2_History.py: Notification history
+- 3_Settings.py: User preferences
+"""
